Moved employee record API serializers to  Django app
