+++ itou/employee_record/serializers.py
@@ -304,56 +304,6 @@ class EmployeeRecordSerializer(serializers.ModelSerializer):
         return result
 
 
-class _API_EmployeeAddress(_EmployeeAddress):
-    # This class in only useful for compatibilty
-    # We decided not to send phone and email (business and bad ASP address filters)
-    # But we make it available for API for compatibility with original document
-    # (these fiels should really be actual data, not fake, by implicit contract)
-
-    def _update_address_and_phone_number(self, result, instance) -> OrderedDict:
-        """
-        Allow overriding these 2 fields:
-        - adrTelephone
-        - adrMail
-        Make data readable again for API users.
-        """
-        result["adrTelephone"] = instance.phone
-        result["adrMail"] = instance.email
-
-        return result
-
-
-class EmployeeRecordAPISerializer(EmployeeRecordSerializer):
-    """
-    This serializer is a version with the `numeroAnnexe` field added (financial annex number).
-
-    This field not needed by ASP was simply ignored in earlier versions of the
-    main SFTP serializer but was removed for RGPD concerns.
-    """
-
-    numeroAnnexe = serializers.CharField(source="financial_annex_number")
-    adresse = _API_EmployeeAddress(source="job_application.job_seeker")
-
-    class Meta:
-        model = EmployeeRecord
-        fields = [
-            "passIae",
-            "passDateDeb",
-            "passDateFin",
-            "numLigne",
-            "typeMouvement",
-            "numeroAnnexe",
-            "mesure",
-            "siret",
-            "personnePhysique",
-            "adresse",
-            "situationSalarie",
-            "codeTraitement",
-            "libelleTraitement",
-        ]
-        read_only_fields = fields
-
-
 class EmployeeRecordBatchSerializer(serializers.Serializer):
     """
     This serializer is a wrapper for a list of employee records
